+++ packages/cli/src/templates/frontend/rest/src/home.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Home() {
-  return <text>Hello Creator, Welcome to your DXP!</text>;
-}