+++ packages/cli/src/util/config/get-package-info.ts
@@ -1,9 +1,12 @@
-import path from "path";
 import fs from "fs-extra";
 import { type PackageJson } from "type-fest";
+import path from "path";
+import { fileURLToPath } from "url";
 
 export function getPackageInfo() {
-  const packageJsonPath = path.join("package.json");
+  const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-  return fs.readJSONSync(packageJsonPath) as PackageJson;
+  const rootPath = path.join(__dirname, "../../../package.json");
+
+  return fs.readJSONSync(rootPath || "") as PackageJson;
 }