+++ services/onboarding/functions/index.ts
@@ -2,7 +2,7 @@ import type { AWS } from "@serverless/typescript";
 
 export const functions: AWS["functions"] = {
   logIn: {
-    handler: "functions/log-in/handler.handler",
+    handler: "functions/log-in.handler",
     events: [
       {
         http: {
@@ -15,7 +15,6 @@ export const functions: AWS["functions"] = {
     environment: {
       SES_FROM_ADDRESS: "noreply@${self:custom.domain}",
       KMS_KEY_ID: { Ref: "EncryptionKey" },
-      BASE_URL: "localhost:5001",
       USER_POOL_ID: { Ref: "PasswordlessMagicLinksUserPool" },
     },
     // @ts-expect-error no types for serverless-iam-roles-per-function
@@ -49,7 +48,7 @@ export const functions: AWS["functions"] = {
     ],
   },
   verify: {
-    handler: "functions/verify/handler.handler",
+    handler: "functions/verify.handler",
     events: [
       {
         http: {
@@ -59,9 +58,12 @@ export const functions: AWS["functions"] = {
         },
       },
     ],
+    environment: {
+      APP_CLIENT_ID: { Ref: "WebUserPoolClient" },
+    },
   },
   logOut: {
-    handler: "functions/log-out/handler.handler",
+    handler: "functions/log-out.handler",
     events: [
       {
         http: {
@@ -73,16 +75,16 @@ export const functions: AWS["functions"] = {
     ],
   },
   preSignUp: {
-    handler: "functions/pre-sign-up/handler.handler",
+    handler: "functions/pre-sign-up.handler",
   },
   defineAuthChallenge: {
-    handler: "functions/define-auth-challenge/handler.handler",
+    handler: "functions/define-auth-challenge.handler",
   },
   createAuthChallenge: {
-    handler: "functions/create-auth-challenge/handler.handler",
+    handler: "functions/create-auth-challenge.handler",
   },
   verifyAuthChallengeResponse: {
-    handler: "functions/verify-auth-challenge-response/handler.handler",
+    handler: "functions/verify-auth-challenge-response.handler",
     environment: {
       KMS_KEY_ID: { Ref: "EncryptionKey" },
     },